requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.2
dnspython==2.4.2
exceptiongroup==1.2.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.339
langsmith==0.0.66
loguru==0.7.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
packaging==23.2
pinecone-client==2.2.4
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
yarl==1.9.3
